public statistics APP side

+++ iSveikata-app/src/main/java/lt/vtvpmc/ems/isveikata/appointment/Appointment.java
@@ -3,12 +3,18 @@ package lt.vtvpmc.ems.isveikata.appointment;
 import java.io.Serializable;
 import java.util.Date;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.OneToOne;
+
+import org.hibernate.annotations.Type;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+
 import lombok.Data;
 import lt.vtvpmc.ems.isveikata.medical_record.MedicalRecord;
-import org.hibernate.annotations.Type;
 
 @Entity
 @Data
@@ -21,7 +27,7 @@ public class Appointment implements Serializable {
 
 	private String description;
 
-	private Long duration;
+	private int duration;
 	
 	@Type(type = "date")
 	private Date date;
